+++ frontend/src/components/ShootDetails/ShootDetails.vue
@@ -8,7 +8,7 @@ SPDX-License-Identifier: Apache-2.0
     <v-row class="d-flex">
       <v-col cols="12" md="6">
         <shoot-details-card :shootItem="shootItem"></shoot-details-card>
-        <shoot-custom-fields-card :shootItem="shootItem" class="mt-4"></shoot-custom-fields-card>
+        <custom-fields-card :custom-fields="customFields" class="mt-4"></custom-fields-card>
         <shoot-infrastructure-card :shootItem="shootItem" class="mt-4"></shoot-infrastructure-card>
         <shoot-external-tools-card :shootItem="shootItem" class="mt-4"></shoot-external-tools-card>
         <shoot-lifecycle-card ref="shootLifecycle" :shootItem="shootItem" class="mt-4"></shoot-lifecycle-card>
@@ -29,10 +29,12 @@ SPDX-License-Identifier: Apache-2.0
 
 <script>
 import { mapGetters } from 'vuex'
+import filter from 'lodash/filter'
 import get from 'lodash/get'
+import map from 'lodash/map'
 
 import ShootDetailsCard from '@/components/ShootDetails/ShootDetailsCard'
-import ShootCustomFieldsCard from '@/components/ShootDetails/ShootCustomFieldsCard'
+import CustomFieldsCard from '@/components/ShootDetails/CustomFieldsCard'
 import ShootExternalToolsCard from '@/components/ShootDetails/ShootExternalToolsCard'
 import ShootInfrastructureCard from '@/components/ShootDetails/ShootInfrastructureCard'
 import ShootLifecycleCard from '@/components/ShootDetails/ShootLifecycleCard'
@@ -49,7 +51,7 @@ export default {
   name: 'shoot-details',
   components: {
     ShootDetailsCard,
-    ShootCustomFieldsCard,
+    CustomFieldsCard,
     ShootInfrastructureCard,
     ShootLifecycleCard,
     ShootAccessCard,
@@ -66,7 +68,8 @@ export default {
   computed: {
     ...mapGetters([
       'ticketsByNamespaceAndName',
-      'canGetSecrets'
+      'canGetSecrets',
+      'customFieldsListShoot'
     ]),
     info () {
       return get(this, 'shootItem.info', {})
@@ -78,6 +81,17 @@ export default {
       const namespace = this.shootNamespace
       const name = this.shootName
       return this.ticketsByNamespaceAndName({ name, namespace })
+    },
+    customFields () {
+      const customFields = filter(this.customFieldsListShoot, ['showDetails', true])
+      return map(customFields, ({ name, path, icon, tooltip, defaultValue }) => ({
+        name,
+        path,
+        icon,
+        tooltip,
+        defaultValue,
+        value: get(this.shootItem, path)
+      }))
     }
   },
   methods: {
